+++ src/data/dao/PurchaseInvoiceDAO.java
@@ -1,12 +1,12 @@
 package data.dao;
 
-import business.entities.Invoice;
+import business.entities.PurchaseInvoice;
 
 /**
  * Created by raoman on 02/10/2016.
  */
 
-public class PurchaseInvoiceDAO extends GenericDAO<Invoice > {
+public class PurchaseInvoiceDAO extends GenericDAO<PurchaseInvoice> {
     private static final String entityClassName = "business.entities.detailsViews";
     private static final String tableName = "purchaseinvoicetable";
