--- src/presentation/gui/DataContainer.fxml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.layout.Pane?>
-<?import javafx.scene.layout.Region?>
-
-
-<Pane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="576.0" prefWidth="466.0" xmlns="http://javafx.com/javafx/8.0.111" xmlns:fx="http://javafx.com/fxml/1">
-   <children>
-      <Button layoutX="69.0" layoutY="483.0" mnemonicParsing="false" text="Button" />
-      <Button layoutX="332.0" layoutY="483.0" mnemonicParsing="false" text="Button" />
-      <Region layoutX="40.0" layoutY="39.0" prefHeight="423.0" prefWidth="384.0" />
-   </children>
-</Pane>
