--- src/presentation/inventory/InvoiceCapture.java
@@ -1,7 +1,0 @@
-package presentation.inventory;
-
-/**
- * Created by cesar on 13/10/16.
- */
-public class InvoiceCapture {
-}
