+++ src/presentation/window/CustomerWindow.java
@@ -1,9 +1,8 @@
-package presentation.gui.window;
+package presentation.window;
 
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
 import javafx.scene.control.Button;
-import javafx.scene.control.TextArea;
 import javafx.scene.control.TextField;
 
 import java.net.URL;
@@ -12,13 +11,13 @@ import java.util.ResourceBundle;
 /**
  * Created by Luis on 03/12/2016.
  */
-public class ServiceWindow implements Initializable {
+public class CustomerWindow implements Initializable {
 
     @FXML protected Button createButton;
     @FXML protected Button cancelButton;
     @FXML protected TextField nameField;
-    @FXML protected TextArea descriptionField;
-    @FXML protected TextField unitariPriceField;
+    @FXML protected TextField telephoneField;
+    @FXML protected TextField emailField;
 
     @Override
     public void initialize(URL location, ResourceBundle resources) {
