+++ src/presentation/window/ServiceWindow.java
@@ -1,7 +1,5 @@
-package presentation.gui.window;
+package presentation.window;
 
-import javafx.event.ActionEvent;
-import javafx.event.EventHandler;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
 import javafx.scene.control.Button;
@@ -11,29 +9,19 @@ import javafx.scene.control.TextField;
 import java.net.URL;
 import java.util.ResourceBundle;
 
-
 /**
  * Created by Luis on 03/12/2016.
  */
-public class SupplierWindow implements Initializable {
+public class ServiceWindow implements Initializable {
 
     @FXML protected Button createButton;
     @FXML protected Button cancelButton;
     @FXML protected TextField nameField;
-    @FXML protected TextArea addressField;
-    @FXML protected TextField telephoneField;
-    @FXML protected TextField emailField;
+    @FXML protected TextArea descriptionField;
+    @FXML protected TextField unitariPriceField;
 
     @Override
     public void initialize(URL location, ResourceBundle resources) {
-
-    }
-
-    public void cancel(){
-
-    }
-
-    public void create(){
 
     }
 }
